requirements.txt
fastapi==0.115.4
opencv-python==4.10.0.84
pydantic==2.9.2
pymongo==4.10.1
ultralytics==8.3.28
uvicorn==0.32.0
python-dotenv

